+++ packages/adapter/src/snap.ts
@@ -1,18 +1,7 @@
-import { FilecoinSnapApi } from "@chainsafe/filsnap-types";
-import {
-  calculateGasForMessage,
-  configure,
-  exportPrivateKey,
-  getAddress,
-  getBalance,
-  getMessages,
-  getPublicKey,
-  sendMessage,
-  signMessage,
-  signMessageRaw,
-} from "./methods";
+import { AptosSnapApi } from "@keystonehq/aptosnap-types";
+import { configure, connect } from "./methods";
 
-export class MetamaskFilecoinSnap {
+export class MetamaskAptosSnap {
   // snap parameters
   protected readonly snapOrigin: string;
   protected readonly snapId: string;
@@ -23,18 +12,10 @@ export class MetamaskFilecoinSnap {
   }
 
   // eslint-disable-next-line @typescript-eslint/require-await
-  public getFilecoinSnapApi = async (): Promise<FilecoinSnapApi> => {
+  public getAptosSnapApi = async (): Promise<AptosSnapApi> => {
     return {
-      calculateGasForMessage: calculateGasForMessage.bind(this),
       configure: configure.bind(this),
-      exportPrivateKey: exportPrivateKey.bind(this),
-      getAddress: getAddress.bind(this),
-      getBalance: getBalance.bind(this),
-      getMessages: getMessages.bind(this),
-      getPublicKey: getPublicKey.bind(this),
-      sendMessage: sendMessage.bind(this),
-      signMessage: signMessage.bind(this),
-      signMessageRaw: signMessageRaw.bind(this),
+      connect: connect.bind(this),
     };
   };
 }