+++ packages/example/src/components/Account/Account.tsx
@@ -1,25 +1,18 @@
 import React from "react";
 import {Box, Button, Card, CardContent, CardHeader, Divider, Grid, Typography} from '@material-ui/core/';
-import {FilecoinSnapApi} from "@chainsafe/filsnap-types";
+import {AptosSnapApi} from "@keystonehq/aptosnap-types";
 
 export interface AccountProps {
     address: string,
-    publicKey: string,
-    balance: string,
-    balanceChange: boolean,
-    api: FilecoinSnapApi | null
+    authenticationKey: string,
+    // balance: string,
+    // balanceChange: boolean,
+    api: AptosSnapApi | null
 }
 
 export const Account = (props: AccountProps) => {
 
-    const handleExport = async () => {
-        if (props.api) {
-            const privateKey = await props.api.exportPrivateKey();
-            alert(`Your private key: ${privateKey}`);
-        }
-    };
-
-    return (        
+    return (
         <Card>
             <CardHeader title="Account details"/>
             <CardContent>
@@ -30,18 +23,15 @@ export const Account = (props: AccountProps) => {
                         <Divider light/>
                         <Box m={"0.5rem"}/>
                         <Typography variant="h6">PUBLIC KEY:</Typography>
-                        <Typography variant="subtitle2">{props.publicKey}</Typography>
+                        <Typography variant="subtitle2">{props.authenticationKey}</Typography>
                         <Divider light/>
                         <Box m={"0.5rem"}/>
-                        <Typography variant="h6">ACCOUNT BALANCE:</Typography>
-                        {props.balanceChange
-                            ? <Typography variant="subtitle2" ><b>{props.balance}</b></Typography>
-                            : <Typography variant="subtitle2" >{props.balance}</Typography>
-                        }
+                        {/*<Typography variant="h6">ACCOUNT BALANCE:</Typography>*/}
+                        {/*{props.balanceChange*/}
+                        {/*    ? <Typography variant="subtitle2" ><b>{props.balance}</b></Typography>*/}
+                        {/*    : <Typography variant="subtitle2" >{props.balance}</Typography>*/}
+                        {/*}*/}
                     </Grid>
-                </Grid>
-                <Grid container item xs={12} justifyContent="flex-end">
-                    <Button color="secondary" variant={"contained"} onClick={handleExport}>Export private key</Button>
                 </Grid>
             </CardContent>
         </Card>