+++ packages/example/src/containers/Dashboard/Dashboard.tsx
@@ -6,10 +6,7 @@ import {
 import {MetaMaskConnector} from "../MetaMaskConnector/MetaMaskConnector";
 import {MetaMaskContext} from "../../context/metamask";
 import {Account} from "../../components/Account/Account";
-import {FilecoinSnapApi, MessageStatus} from "@chainsafe/filsnap-types";
-import {TransactionTable} from "../../components/TransactionTable/TransactionTable";
-import {SignMessage} from "../../components/SignMessage/SignMessage";
-import {Transfer} from "../../components/Transfer/Transfer";
+import {AptosSnapApi} from "@keystonehq/aptosnap-types";
 import Footer from "../../Footer";
 
 export const Dashboard = () => {
@@ -17,96 +14,84 @@ export const Dashboard = () => {
     const [state] = useContext(MetaMaskContext);
     const [balance, setBalance] = useState("");
     const [address, setAddress] = useState("");
-    const [publicKey, setPublicKey] = useState("");
-    const [messages, setMessages] = useState<MessageStatus[]>([]);
-    const [networks, setNetworks] =  useState<"t"|"f">("f")
+    const [authenticationKey, setAuthenticationKey] = useState("");
+    const [networks, setNetworks] =  useState<"mainnet"|"devnet">("devnet")
 
-    const [balanceChange, setBalanceChange] = useState<boolean>(false);
+    const [network, setNetwork] = useState<"mainnet" | "devnet" >("devnet");
 
-    const [network, setNetwork] = useState<"f" | "t" >("f");
-
-    const [api, setApi] = useState<FilecoinSnapApi|null>(null);
+    const [api, setApi] = useState<AptosSnapApi|null>(null);
 
     const handleNetworkChange = async (event: React.ChangeEvent<{value: any}>) => {
-        const selectedNetwork = event.target.value as "f" | "t";
+        const selectedNetwork = event.target.value as "mainnet" | "devnet";
         if (selectedNetwork === network) return;
         if (api) {
             try {
                 await api.configure({network: selectedNetwork});
                 setNetworks(selectedNetwork)
                 setNetwork(selectedNetwork);
-                setMessages(await api.getMessages());
             } catch(e) {
                 console.error("Unable to change network", e)
             }
         }
     };
 
-    const handleNewMessage = useCallback(async () => {
-        if (api) {
-            setMessages(await api.getMessages());
-        }
-    }, [api, setMessages]);
-
     useEffect(() => {
         (async () => {
-            if (state.filecoinSnap.isInstalled && state.filecoinSnap.snap) {
-                const filecoinApi = await state.filecoinSnap.snap.getFilecoinSnapApi();
-                setApi(filecoinApi);
+            if (state.aptosSnap.isInstalled && state.aptosSnap.snap) {
+                const aptosApi = await state.aptosSnap.snap.getAptosSnapApi();
+                console.log("aptosApi", aptosApi);
+                setApi(aptosApi);
             }
         })();
-    }, [state.filecoinSnap.isInstalled, state.filecoinSnap.snap]);
+    }, [state.aptosSnap.isInstalled, state.aptosSnap.snap]);
 
     useEffect(() => {
         (async () => {
             if (api) {
-                setAddress(await api.getAddress());
-                setPublicKey(await api.getPublicKey());
-                setBalance(await api.getBalance());
-                setMessages(await api.getMessages());
+                const account = await api.connect();
+                setAddress(account.address().hex());
+                setAuthenticationKey(account.authKey().toString());
+                // setBalance(await api.getBalance());
             }
         })();
     }, [api, network]);
 
-    useEffect( () => {
-        // periodically check balance
-        const interval = setInterval(async () => {
-            if (api) {
-                const newBalance = await api.getBalance();
-                if (newBalance !== balance) {
-                    setBalanceChange(true);
-                    setBalance(newBalance);
-                } else {
-                    setBalanceChange(false)
-                }
-            }
-        }, 30000); // every 30 seconds ~ 1 epoch
-        return () => clearInterval(interval);
-    }, [api, balance, setBalance, setBalanceChange]);
+    // useEffect( () => {
+    //     // periodically check balance
+    //     const interval = setInterval(async () => {
+    //         if (api) {
+    //             const newBalance = await api.getBalance();
+    //             if (newBalance !== balance) {
+    //                 setBalanceChange(true);
+    //                 setBalance(newBalance);
+    //             } else {
+    //                 setBalanceChange(false)
+    //             }
+    //         }
+    //     }, 30000); // every 30 seconds ~ 1 epoch
+    //     return () => clearInterval(interval);
+    // }, [api, balance, setBalance, setBalanceChange]);
 
     return (
         <Container maxWidth="lg">
             <Grid direction="column" alignItems="center" justifyContent="center" container spacing={3}>
                 <Box m="2rem" style={{textAlign: "center"}}>
                     <Typography variant="h2">
-                        Filsnap demo
-                    </Typography>
-                    <Typography style={{color: "gray", fontStyle: "italic"}} variant="h6">
-                        Filsnap enables Filecoin network inside Metamask.
+                        Aptos demo
                     </Typography>
                 </Box>
-                <Hidden xsUp={state.filecoinSnap.isInstalled}>
+                <Hidden xsUp={state.aptosSnap.isInstalled}>
                     <MetaMaskConnector/>
                     <Footer/>
                 </Hidden>
-                <Hidden xsUp={!state.filecoinSnap.isInstalled}>
+                <Hidden xsUp={!state.aptosSnap.isInstalled}>
                     <Box m="1rem" alignSelf="baseline">
                         <InputLabel>Network</InputLabel>
                         <Select
                             onChange={handleNetworkChange}
                             value={networks}
                         >
-                            <MenuItem value={"t"}>Testnet</MenuItem>
+                            <MenuItem value={"t"}>Devnet</MenuItem>
                             <MenuItem value={"f"}>Mainnet</MenuItem>
                         </Select>
                     </Box>
@@ -114,31 +99,11 @@ export const Dashboard = () => {
                         <Grid item xs={12}>
                             <Account
                                 address={address}
-                                balance={balance + " FIL"}
-                                publicKey={publicKey}
+                                // balance={balance + " FIL"}
+                                authenticationKey={authenticationKey}
                                 api={api}
-                                balanceChange={balanceChange}
+                                // balanceChange={balanceChange}
                             />
-                        </Grid>
-                    </Grid>
-                    <Box m="1rem"/>
-                    <Grid container spacing={3} alignItems="stretch">
-                        <Grid item md={6} xs={12}>
-                            <Transfer api={api} network={network} onNewMessageCallback={handleNewMessage} />
-                        </Grid>
-                        <Grid item md={6} xs={12}>
-                            <SignMessage api={api} />
-                        </Grid>
-                    </Grid>
-                    <Box m="1rem"/>
-                    <Grid container spacing={3} alignItems={"stretch"}>
-                        <Grid item xs={12}>
-                            <Card>
-                                <CardHeader title="Account transactions"/>
-                                <CardContent>
-                                    <TransactionTable txs={messages}/>
-                                </CardContent>
-                            </Card>
                         </Grid>
                     </Grid>
                 </Hidden>