+++ packages/example/src/services/metamask.ts
@@ -1,5 +1,5 @@
-import {SnapRpcMethodRequest} from "@chainsafe/filsnap-types";
-import {enableFilecoinSnap, MetamaskFilecoinSnap} from "@chainsafe/filsnap-adapter";
+import {SnapRpcMethodRequest} from "@keystonehq/aptosnap-types";
+import {enableAptosSnap, MetamaskAptosSnap} from "@keystonehq/aptosnap-adapter";
 
 declare global {
     interface Window {
@@ -19,24 +19,19 @@ let isInstalled: boolean = false;
 
 export interface SnapInitializationResponse {
     isSnapInstalled: boolean;
-    snap?: MetamaskFilecoinSnap;
+    snap?: MetamaskAptosSnap;
 }
 
-export async function initiateFilecoinSnap(): Promise<SnapInitializationResponse> {
+export async function initiateAptosSnap(): Promise<SnapInitializationResponse> {
     const snapId = process.env.REACT_APP_SNAP_ID ? process.env.REACT_APP_SNAP_ID : defaultSnapId
     try {
-        console.log('Attempting to connect to snap...');
-        const metamaskFilecoinSnap = await enableFilecoinSnap({network: "f"}, snapId, {version: "latest"});
+        const metamaskAptosSnap = await enableAptosSnap({network: "devnet"}, snapId, {version: "latest"});
         isInstalled = true;
         console.log('Snap installed!');
-        return {isSnapInstalled: true, snap: metamaskFilecoinSnap};
+        return {isSnapInstalled: true, snap: metamaskAptosSnap};
     } catch (e) {
         console.error(e);
         isInstalled = false;
         return {isSnapInstalled: false};
     }
-}
-
-export async function isFilecoinSnapInstalled(): Promise<boolean> {
-    return isInstalled;
 }